+++ BuscoCasa/Controllers/bookmarks/BookmarksViewController.swift
@@ -28,18 +28,7 @@ class BookmarksViewController: UIViewController {
         
 
         let retrievedUser: String? = KeychainWrapper.standard.string(forKey: AppConstants.UserConstants.userSaveData)
-        if retrievedUser != nil {
-            if let jsonData = retrievedUser!.data(using: .utf8)
-            {
-                let decoder = JSONDecoder()
-                
-                do {
-                    self.user = try decoder.decode(User.self, from: jsonData)
-                } catch {
-                    print(error.localizedDescription)
-                }
-            }
-        }
+        self.user = GetLoggedUser.getLoggedUser()
 
         bookmarksTable.dataSource = self
         bookmarksTable.delegate = self
