+++ BuscoCasa/Controllers/user/UserViewController.swift
@@ -75,19 +75,8 @@ class UserViewController: UIViewController, ModalDelegate {
         userPicture.layer.borderWidth = 4
         userPicture.layer.borderColor = UIColor.white.cgColor
         
-        let retrievedUser: String? = KeychainWrapper.standard.string(forKey: AppConstants.UserConstants.userSaveData)
-        if retrievedUser != nil {
-            if let jsonData = retrievedUser!.data(using: .utf8)
-            {
-                let decoder = JSONDecoder()
-                
-                do {
-                    self.user = try decoder.decode(User.self, from: jsonData)
-                } catch {
-                    print(error.localizedDescription)
-                }
-            }
-        }
+        self.user = GetLoggedUser.getLoggedUser()
+
         setUserData()
     }
